requirements.txt
Django==1.10.2
django-filter==1.0.1
django-model-utils==2.6
djangorestframework==3.5.3
flake8==3.2.1
mccabe==0.5.3
pycodestyle==2.2.0
pyflakes==1.3.0
requests==2.12.4
vulture==0.12
